fix: without pha-worker (#569)
* fix: without pha-worker

* chore: optimize code
